Edit: Separate notfound component + CSS module into folder
